PBMS-64: Implement Navbar that has Logo and tabs Navbar is complete per Jira description, minor css errors may occur later in the project and will be fixed accordingly.
In-Progress:
PBMS-69: Add footer to website
Implemented footer to the website to be viewed across all pages. Footer is not done, but the general layout is good to go.

Modified App.jsx for Navbar and Footer to be viewed across pages.

Modified testimonials.jsx for minor error as the footer overlapped the bottom image. Now fixed.