PBMS-115
Dug deep into the UI of the home.jsx to adjust scaling of text, images, etc when viewing from different devices. Also general styling adjustments/fixes.

App.jsx: discrepancy with import for signup page, was causing crash (white page) while running server.

index.css: Added color palette and font-family of client's choosing. Currently, the color palette is commented out so it does not affect the current design of the site until we are ready to do so. font-serif and font-mono use the same font 'Bodoni Moda', I created a variable so this reflects ALL the areas where -mono or -serif is used in the codebase.

Footer.jsx: adjusted to reflect device screen sizes

Navbar.jsx: Adjusted Navbar to tranisiton into a hamburger menu when the screen size is small, specifically on mobile devices. When the hamburger menu is clicked, it will pop open a sidebar with the tabs to the other pages. Also move the special services link so its position reflects being next to the services tab.

Added .png files of the client's logo in the /public folder and deleted unecessary .svg files.